Branch yzl (#2)
* 新增 token 过滤器

* 新增 token 过滤器

+++ src/main/java/com/jerry/savior_oauth/pojo/BO/UserInfoBO.java
@@ -40,6 +40,6 @@ public class UserInfoBO {
 
     private Date createTime;
 
-    private Set<Integer> permissionSet;
+    private Set<Long> permissionSet;
 
 }
